Replace `base_url` with `api_url` in the client constructor
On the platform, we have only the API server URL without the version suffix available as environment variables in the actor,
so you can't supply them directly to the client. This splits it so that users can supply just the API url to the client constructor,
and the API version suffix gets appended internally (the utilized API version is an internal thing of this client,
it shouldn't be passed from the outside in the first place).