Improve retrying failed API server connections (#71)
When the API server got terminated in the middle of a connection (perhaps because of a new deployment), it didn't finish the response properly, and the connection errored out with a `"Connection broken: InvalidChunkLength(got length b'', 0 bytes read)"` error. This change extends the error-retrying logic to retry the request in case of those errors as well.