Unified methods for streaming resources (#80)
With the implementation of the Apify SDK for Python, we will have to do
some bigger changes to the Python Client, so now is a good time to take
some time and get rid of some old problems, before we are stuck with
them forever. One of them is that streaming methods for some resources
(dataset items, key-value store records and log contents) are a mess.

This unifies their behavior, and makes them safer by forcing them to
return a context-managed object, so the streaming response is
automatically closed after use.

It also unifies the naming and behavior of methods for fetching the
resource contents as raw bytes, by naming them all in the `get_as_bytes`
style, instead of the varying names and usages they had before.