Switched underlying HTTP library from `requests` to `httpx` (#81)
We want to introduce `async` mode to the client, so that users can make
non-blocking requests using `asyncio`. To achieve that, we have to
replace the underlying HTTP library, since `requests` does not have
`async` support.

I'm switching it to [`httpx`](https://github.com/encode/httpx/) here.
It's a fairly mature HTTP(S) library, with support for both sync and
async operation, and all the features we need, and it has a (mostly)
`requests`-compatible interface, so the switch is fairly easy.

There should be basically no functional differences, but still I'm
marking the change as breaking, since under the hood it uses `httpcore`
on the transport layer, but `requests` uses `urllib3`, so there could be
some low level changes which could possibly affect someone.