Add `ignore_docs` decorator (#112)
This adds an `ignore_docs` decorator, which does not do anything on its
own, it just serves as a mark that we will use in the new documentation
rendering to mark methods that we want to ignore there.

I've added it to all the `__init__` methods of all the classes except
for `ApifyClient` and `ApifyClientAsync`, because those are the only two
classes that we want users to actually create. All the other classes
should not be constructed externally.