refactor: Rework client timeout (#384)
Add optional timeout argument to the individual client calls. Timeout of
each call is either this timeout or the default client timeout.
In case of failed, but retry-able request, the timeout is multiplied by
2^attempt in each consecutive attempt, up to the value of client default
timeout.

Several storage related endpoints now have reduced timeout value. Mainly
the endpoints that are not expected to take long or are considered safe
to run twice or idempotent.

Timeouts for the storage related endpoints are all summarized in the new
tests where it is all in one place.

Addresses issues found in:
https://github.com/apify/crawlee-python/issues/1132